chore: Refactor reportModel.ts
