+++ src/main/model/adapters/report/detectedLicenseSummaryAdapter.ts
